Add color coding to log messages
This commit adds color coding to log messages based on their level. Debug messages are grey, warnings are yellow, and errors are red. The code uses a switch statement to determine the color based on the log level. The commit also includes a commented out line of code that would have localized the log message timestamp, but this feature is not currently used.