Add environment variable to log output
This commit adds an environment variable to the log output. If the environment variable ENVIRONMENT is not set to production, the log output will include a full caller path. If it is set to production, the log output will include an environment field with the value of the ENVIRONMENT variable. The commit also removes the time from the console log output.