Update clog dependency in go.mod
This commit updates the clog dependency in the go.mod file. The previous version was v0.0.0-20230929023203-6b1128fc5ab3, and it has been updated to v0.0.0-20230930052527-683e18a3d08d. This change was made to ensure compatibility and use the latest version of clog.

Additionally, the go.sum file has been updated to reflect this change. The previous hash for the clog dependency was h1:nadS2BV4kLVNPvqbsbgOWF9ykDeBpLCTdpjSK0xFNpg=, and it has been updated to h1:dIVVSvq1NzQMf/eYTYZiPRy5D1kErYn++8bND9rdEyU=. This change ensures that the correct version of clog is used when building the project.

No other dependencies or files were affected by this commit.